Blog/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-magic==0.4.27
drf-yasg==1.21.7